Add task controller

+++ telegram-bot/src/main/java/ru/picker/core/service/ChapterService.java
@@ -5,11 +5,14 @@ import org.springframework.stereotype.Service;
 import ru.picker.core.entity.Chapter;
 import ru.picker.core.entity.SubChapter;
 import ru.picker.core.mapper.ChapterMapper;
+import ru.picker.core.model.ChapterDisplayDto;
 import ru.picker.core.model.IncomeChapterDto;
 import ru.picker.core.repository.ChapterRepository;
 
 import javax.ws.rs.NotFoundException;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.UUID;
 
 @Service
 @RequiredArgsConstructor
@@ -17,17 +20,19 @@ public class ChapterService {
 
     private final ChapterRepository chapterRepository;
 
-    public Chapter add(IncomeChapterDto incomeChapterDto) {
-        return chapterRepository.save(ChapterMapper.INSTANCE.map(incomeChapterDto));
+    public ChapterDisplayDto add(IncomeChapterDto incomeChapterDto) {
+        Chapter chapter = ChapterMapper.INSTANCE.map(incomeChapterDto);
+        chapterRepository.save(chapter);
+        return ChapterMapper.INSTANCE.map(chapter);
     }
 
-    public Chapter findById(UUID id) {
-        return chapterRepository.findById(id).orElseThrow(() ->
-                new NotFoundException(String.format("Chapter with ID: %s not found", id)));
+    public ChapterDisplayDto get(String id) {
+        Chapter chapter = findById(id);
+        return ChapterMapper.INSTANCE.map(chapter);
     }
 
-    public Set<Chapter> getAllChapters() {
-        return new HashSet<>(chapterRepository.findAll());
+    public List<Chapter> getAllChapters() {
+        return chapterRepository.findAll();
     }
 
     public List<SubChapter> getSubChaptersByTheme(String theme) {
@@ -37,19 +42,27 @@ public class ChapterService {
         return new ArrayList<>(chapter.getSubChapters());
     }
 
-    public Chapter renewChapter(UUID id, IncomeChapterDto incomeChapterDto) {
+    public ChapterDisplayDto renewChapter(String id, IncomeChapterDto incomeChapterDto) {
         Chapter chapter = findById(id);
         if (incomeChapterDto.getName() != null && !incomeChapterDto.getName().isBlank()) {
             chapter.setName(incomeChapterDto.getName());
             chapterRepository.save(chapter);
         }
-        return chapter;
+        return ChapterMapper.INSTANCE.map(chapter);
     }
 
-    public void deleteChapter(UUID id) {
+    public void deleteChapter(String id) {
         Chapter chapter = findById(id);
-        chapterRepository.delete(chapter);
+        if (chapter != null) {
+            chapterRepository.delete(chapter);
+        }
     }
+
+    public Chapter findById(String id) {
+        return chapterRepository.findById(UUID.fromString(id)).orElseThrow(() ->
+                new NotFoundException(String.format("Chapter with ID: %s not found", id)));
+    }
+
 }
