+++ telegram-bot/src/main/java/ru/picker/core/service/SubChapterService.java
@@ -2,12 +2,11 @@ package ru.picker.core.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import ru.picker.core.entity.Chapter;
 import ru.picker.core.entity.SubChapter;
 import ru.picker.core.entity.Task;
 import ru.picker.core.mapper.SubChapterMapper;
 import ru.picker.core.model.IncomeSubChapterDto;
-import ru.picker.core.model.SubChapterDto;
+import ru.picker.core.model.SubChapterDisplayDto;
 import ru.picker.core.repository.SubChapterRepository;
 
 import javax.ws.rs.NotFoundException;
@@ -21,22 +20,20 @@ import java.util.UUID;
 public class SubChapterService {
 
     private final SubChapterRepository subChapterRepository;
-    private final ChapterService chapterService;
-    private final TaskService taskService;
 
-    public SubChapter add(IncomeSubChapterDto incomeSubChapterDto) {
-        return subChapterRepository.save(SubChapterMapper.INSTANCE.map(
-                incomeSubChapterDto,
-                chapterService));
+    public SubChapterDisplayDto add(IncomeSubChapterDto incomeSubChapterDto) {
+        SubChapter subChapter = SubChapterMapper.INSTANCE.map(incomeSubChapterDto);
+
+        return SubChapterMapper.INSTANCE.map(subChapterRepository.save(subChapter));
     }
 
-    public SubChapter findById(UUID id) {
-        return subChapterRepository.findById(id).orElseThrow(() ->
-                new NotFoundException(String.format("SubChapter with ID: %s not found", id)));
+    public SubChapterDisplayDto get(String id) {
+        SubChapter subChapter = findById(id);
+        return SubChapterMapper.INSTANCE.map(subChapter);
     }
 
-    public Set<SubChapter> findAllByChapterId(UUID chapterId) {
-        return subChapterRepository.findAllByChapter_Id(chapterId);
+    public Set<SubChapter> findAllByChapterId(String chapterId) {
+        return subChapterRepository.findAllByChapter_Id(UUID.fromString(chapterId));
     }
 
     public List<Task> getAllTasksBySubChapter(String theme) {
@@ -50,21 +47,28 @@ public class SubChapterService {
         return subChapterRepository.findAll();
     }
 
-    public SubChapterDto renewSubChapter(UUID id, IncomeSubChapterDto incomeSubChapterDto) {
+    public SubChapterDisplayDto renewSubChapter(String id, IncomeSubChapterDto incomeSubChapterDto) {
         SubChapter subChapter = findById(id);
         if (incomeSubChapterDto.getName() != null && !incomeSubChapterDto.getName().isBlank()) {
             subChapter.setName(incomeSubChapterDto.getName());
         }
-        if (incomeSubChapterDto.getChapterId() != null) {
-            Chapter chapter = chapterService.findById(incomeSubChapterDto.getChapterId());
-            subChapter.setChapter(chapter);
+        if (incomeSubChapterDto.getChapter() != null) {
+            subChapter.setChapter(incomeSubChapterDto.getChapter());
         }
         subChapterRepository.save(subChapter);
-        return SubChapterMapper.INSTANCE.map(subChapter, taskService);
+        return SubChapterMapper.INSTANCE.map(subChapter);
     }
 
-    public void deleteSubChapter(UUID id) {
+    public void deleteSubChapter(String id) {
         SubChapter subChapter = findById(id);
-        subChapterRepository.delete(subChapter);
+        if (subChapter != null) {
+            subChapterRepository.delete(subChapter);
+        }
+    }
+
+    public SubChapter findById(String id) {
+
+        return subChapterRepository.findById(UUID.fromString(id)).orElseThrow(() ->
+                new NotFoundException(String.format("SubChapter with ID: %s not found", id)));
     }
 }
